Fully implemented Depth First Search
Main - Created a temporary map and a DFS object to test the algorithm and output

PathFinder - Added a method to display the map, and a method to create the final path into a vector array. Also adjusted the setBoardMap to update the board given x and y coordinates.

DepthFirstSearch - Fully implemented the DFS algorithm into the findPath virtual method along with validation of the discovered coordinates and whether a path has been found or not.